Add support for posting notices as GitLab Issues
Slack is great, but we prefer track events like this as issues to follow up on
via a private GitLab server. If the right combination of GITLAB env
variables are set, this will create a new issue for each new release.

Since this adds a new destination type, and it's uncertain if that is
desirable, for now there is bit of code duplication just to demonstrate
the feature.

Thanks for considering!